fix: modals break with z-index on main
this removes negative margin. Previously we would tuck the footer under the main section, which would have a z-index and a shadow.

But, z-index on main messes with bootstrap 4's modals. So... instead this adds an inset shadow and removes the tuck so we can remove the z-index.

Probably a cleaner solution anyway.