fix dynamic breakpoints
useBreakpoints still grumbles about `onMounted`. it seems that useBreakpoints actually creates a new composable to be used for the given breakpoint. Since our breakpoints might change after a component is mounted, this watches for the change and then creates a new composable for the new breakpoint.

Feels hacky, but it works.