rework atBreakpoint
This simplifies a bit. It changes to use a computed value rather than a watch, which still allows us to avoid the onMounted warning. The trick is to use the `greaterOrEqual` ref rather than `isGreaterOrEqual`. `isGreaterOrEqual` returns a raw boolean, and so the value of `atBreakpoint` won't be computed. Using the ref will let us recompute the atBreakpoint value and will handle menuBreakpoint prop changes.

Also, this updates child components to use the injection key.