fix: Navbar doesn't work as standalone component
In release `1.1.0` we introduced better breakpoint detection to handle navbars with a lot of options (like Bluesheet) – those navbars run out of screen real estate faster, and should switch to the collapsible mobile nav sooner.

Because a number of different components needed to use the same breakpoint property info (menu button, mobile nav, dropdown, navbar) This was implemented using provide/injects to avoid prop drilling. The `<AppHeader>` provided the boolean ref `atBreakpoint` to any children.

This prevents using `<Navbar>` as a standalone component, as it won't have access to the provided `atBreakpoint` value if it's not within an `<AppHeader>`

The fix is to give the `inject` function a fallback breakpoint value `always`, meaning it the navbar will always show if a breakpoint is not defined.